Raise ordinance on BuildFlowAction
We want to ensure we are close to the top of the page when rendering
the flow.

Note: There is no real documentation on good values to use for
"ordinal". Some plugins aim for low values (1-10). In core, some
ordinals are defined in steps of 100s (subtracted from MAX_INT).

Change-Id: I645c977d2591522be4f0c80a3a397470caf36b58